load_bsa: SSS RRR
We need to pass load_cache=True via __init__ (or do_update) to init the
self._assets variable. I removed the class variable but note handling in
IniFIle - it's still a bit weird - again the root issue is encapsulate
__init__ - we want all initializations to be with load_cache=True - but
note that anyway currently a do_update that fails won't reset the
assets (but will (correctly) reset _ci_settings_cache_linenum).
So let's have a new param for load_bsa - be explicit here. We need
to encapsulate __init__ (to the point of a _in_factory param that
will raise if False in ListInfo init) but that passes through abs_path
handling - ghost of # 336